editor: Move line piece table storage
The editor module exposes cursor movement, line selection, generated line
insertion, and buffer export through Editor.

Its storage layer also lives in edit.py, so the public editing facade sits
beside the run-table code that tracks borrowed line ranges. That makes the
module boundary harder to follow when changing editor behavior.

This commit addresses that by moving the line piece-table records and
run-table implementation into editor.piece_table while leaving Editor as
the facade over those operations.

The following tests commit will add a boundary check for the extracted
editor storage module.
